backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
requests>=2.31.0
python-dotenv>=1.0.1
telethon>=1.36.0
websockets>=12.0

